requirements.txt
absl-py==2.0.0
accelerate==0.23.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
basemap==1.3.8
basemap-data==1.3.2
basicsr==1.4.2
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.9.1
bleach==6.0.0
cachetools==5.3.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.83
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.5.2
cmake==3.27.5
comm==0.1.4
compel==0.1.8
contourpy==1.1.1
cryptography==36.0.2
cycler==0.12.0
dataclasses==0.6
datasets==2.14.5
dctorch==0.1.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
-e git+https://github.com/huggingface/diffusers@2e53936c97d167713c9e97414160124861fa4b68#egg=diffusers
dill==0.3.7
docker-pycreds==0.4.0
docutils==0.19
einops==0.6.1
etils==1.5.0
exceptiongroup==1.1.3
execnet==2.0.2
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
flatbuffers==23.5.26
flax==0.7.4
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
ftfy==6.1.1
funcx-common==0.0.15
future==0.18.3
fuzzywuzzy==0.18.0
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.37
globus-cli==3.4.0
globus-sdk==3.5.0
google==3.0.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gql==3.4.1
graphql-core==3.2.3
grpcio==1.59.0
h5py==3.9.0
hf-doc-builder==0.4.0
huggingface-hub==0.17.3
idna==3.6
imageio==2.31.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
install==1.3.5
invisible-watermark==0.2.0
ipykernel==6.25.2
ipython==8.16.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
jax==0.4.16
jaxlib==0.4.16
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonmerge==1.9.2
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
k-diffusion==0.1.0
keras==2.14.0
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
lightning-utilities==0.9.0
lit==17.0.1
llvmlite==0.41.0
lmdb==1.4.1
lockfile==0.12.2
lpips==0.1.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
mpi4py==3.1.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.4
notebook_shim==0.2.3
numba==0.58.0
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.1
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parameterized==0.9.0
paramiko==2.11.0
parsl==2022.10.10
parso==0.8.3
pathspec==0.11.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.0
Pillow==10.1.0
platformdirs==3.10.0
pluggy==1.3.0
pooch==1.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.16.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyproj==3.6.1
pyshp==2.3.1
PyTDC==0.4.1
pytest==7.4.2
pytest-timeout==2.1.0
pytest-xdist==3.3.1
python-daemon==2.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-fid==0.3.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
rdkit-pypi==2022.9.5
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-mock==1.10.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rotary-embedding-torch==0.3.0
rpds-py==0.10.3
rsa==4.9
ruff==0.0.291
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.6
stack-data==0.6.2
sympy==1.12
tblib==1.7.0
tensorboard==2.16.2
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorstore==0.1.45
termcolor==2.3.0
terminado==0.17.1
texttable==1.6.4
threadpoolctl==3.2.0
tifffile==2023.9.26
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
torch-fidelity==0.3.0
torch_geometric==2.4.0
torchdiffeq==0.2.3
torchmetrics==1.1.2
torchsde==0.2.6
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
trampoline==0.1.2
transformers==4.33.3
triton==2.1.0
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wandb==0.15.11
wcwidth==0.2.7
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==10.3
Werkzeug==2.3.7
widgetsnbextension==4.0.9
wrapt==1.14.1
xxhash==3.3.0
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
